automations/invoice-processor/requirements.txt
pdfplumber==0.10.3
python-dotenv==1.0.0
requests==2.31.0
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
